after much cleanup, I synced the changes from the upstream repo
and rebased my changes on top of it. I had to resolve a few conflicts,
but now I'm ready to commit the changes.
The colorschemes are mostly gone, I need to check out earlier commits to
find them. I should also move things to the lua user or custom folders
to keep things organized.